+++ src/i_video.c
@@ -901,25 +901,39 @@ void I_UpdateBlitFunc(const bool shaking)
     }
 }
 
+static byte     *gamma;
+static float    saturation;
+static float    contrast;
+static float    red;
+static float    green;
+static float    blue;
+
+void I_UpdateColors(void)
+{
+    gamma = gammatable[gammaindex];
+    saturation = r_saturation / 100.0f;
+    contrast = (259.0f * (r_contrast + 255.0f)) / (255.0f * (259.0f - r_contrast));
+    red = r_red / 100.0f;
+    green = r_green / 100.0f;
+    blue = r_blue / 100.0f;
+
+    I_SetPalette(&PLAYPAL[st_palette * 768]);
+
+    if (!vid_pillarboxes)
+        SDL_SetRenderDrawColor(renderer, 0, 0, 0, SDL_ALPHA_OPAQUE);
+}
+
 //
 // I_SetPalette
 //
 void I_SetPalette(const byte *playpal)
 {
-    const byte  *gamma = gammatable[gammaindex];
-    const float saturation = r_saturation / 100.0f;
-    const float contrast = (259.0f * (r_contrast + 255.0f)) / (255.0f * (259.0f - r_contrast));
-
     for (int i = 0; i < 256; i++)
     {
-        byte        r = (byte)(gamma[*playpal++] * r_red / 100.0f);
-        byte        g = (byte)(gamma[*playpal++] * r_green / 100.0f);
-        byte        b = (byte)(gamma[*playpal++] * r_blue / 100.0f);
+        byte        r = (byte)(gamma[*playpal++] * red);
+        byte        g = (byte)(gamma[*playpal++] * green);
+        byte        b = (byte)(gamma[*playpal++] * blue);
         const float p = saturationtable[r][g][b];
-
-        r = (byte)BETWEENF(0.0f, r * (r_red / 100.0f), 255.0f);
-        g = (byte)BETWEENF(0.0f, g * (r_green / 100.0f), 255.0f);
-        b = (byte)BETWEENF(0.0f, b * (r_blue / 100.0f), 255.0f);
 
         r = (byte)BETWEENF(0.0f, p + (r - p) * saturation, 255.0f);
         g = (byte)BETWEENF(0.0f, p + (g - p) * saturation, 255.0f);
@@ -1726,6 +1740,7 @@ static void SetVideoMode(const bool createwindow, const bool output)
     if (SDL_SetSurfacePalette(surface, palette) < 0)
         I_SDLError("SDL_SetSurfacePalette", -1);
 
+    I_UpdateColors();
     I_SetPalette(&PLAYPAL[st_palette * 768]);
 
     src_rect.w = SCREENWIDTH;
@@ -1887,14 +1902,6 @@ void I_ToggleFullscreen(const bool output)
 
         PositionOnCurrentDisplay();
     }
-}
-
-void I_SetPillarboxes(void)
-{
-    I_SetPalette(&PLAYPAL[st_palette * 768]);
-
-    if (!vid_pillarboxes)
-        SDL_SetRenderDrawColor(renderer, 0, 0, 0, SDL_ALPHA_OPAQUE);
 }
 
 static void I_InitPaletteTables(void)